+++ tests/core/test_calculators.py
@@ -31,7 +31,7 @@ class TestPixelStats:
         assert "histogram" in result["stats"]
 
         assert len(result["stats"]["mean"]) == 1
-        assert result["stats"]["mean"].dtype == np.float16
+        assert result["stats"]["mean"].dtype == np.float32
         assert len(result["stats"]["histogram"][0]) == 256
 
     def test_process_with_nans(self):
@@ -192,7 +192,7 @@ class TestVisualStats:
         assert "percentiles" in result["stats"]
 
         assert len(result["stats"]["brightness"]) == 1
-        assert result["stats"]["brightness"].dtype == np.float16
+        assert result["stats"]["brightness"].dtype == np.float32
         assert len(result["stats"]["percentiles"][0]) == 5  # QUARTILES length
 
 
@@ -638,7 +638,7 @@ class TestLowerDimensionalPixelStats:
         assert "histogram" in result["stats"]
 
         assert len(result["stats"]["mean"]) == 1
-        assert result["stats"]["mean"].dtype == np.float16
+        assert result["stats"]["mean"].dtype == np.float32
         assert len(result["stats"]["histogram"][0]) == 256
 
     def test_2d_data_pixel_stats(self):
@@ -659,7 +659,7 @@ class TestLowerDimensionalPixelStats:
         assert "histogram" in result["stats"]
 
         assert len(result["stats"]["mean"]) == 1
-        assert result["stats"]["mean"].dtype == np.float16
+        assert result["stats"]["mean"].dtype == np.float32
         assert len(result["stats"]["histogram"][0]) == 256
 
     def test_1d_data_with_nans(self):
@@ -708,7 +708,7 @@ class TestLowerDimensionalVisualStats:
         assert "percentiles" in result["stats"]
 
         assert len(result["stats"]["brightness"]) == 1
-        assert result["stats"]["brightness"].dtype == np.float16
+        assert result["stats"]["brightness"].dtype == np.float32
         assert len(result["stats"]["percentiles"][0]) == 5  # QUARTILES length
         # Sharpness should be NaN for 1D data
         assert np.isnan(result["stats"]["sharpness"][0])
@@ -726,10 +726,10 @@ class TestLowerDimensionalVisualStats:
         assert "percentiles" in result["stats"]
 
         assert len(result["stats"]["brightness"]) == 1
-        assert result["stats"]["brightness"].dtype == np.float16
+        assert result["stats"]["brightness"].dtype == np.float32
         assert len(result["stats"]["percentiles"][0]) == 5
         # Sharpness should be computed for 2D data
-        assert result["stats"]["sharpness"].dtype == np.float16
+        assert result["stats"]["sharpness"].dtype == np.float32
         assert not np.isnan(result["stats"]["sharpness"][0])
 
     def test_1d_data_visual_stats_per_channel(self):
